Share Paseo Agent env reference helpers
